requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.1.8
django-cors-headers==3.14.0
django-mysql==4.9.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
inflection==0.5.1
netifaces==0.11.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
whitenoise==6.4.0
